Rename mods package to modes
