+++ src/main/java/ch/epfl/planair/modes/PlayMode.java
@@ -1,4 +1,4 @@
-package ch.epfl.planair.mods;
+package ch.epfl.planair.modes;
 
 import ch.epfl.planair.Planair;
 import ch.epfl.planair.meta.Constants;
